updating App.js and TodoList.js
I was looking at the Code Sandbox from earlier today and thinking that maybe if I tried putting my stuff in the App.js instead of TodoList.js, I could more easily compare between what we did in class today and what I'm trying to accomplish now. It's not a 1:1 comparison or anything but I think I was kind of doing it backwards before.