Complete basic todo list
I went through and completed the MVP for day one of the ToDo List project. It was a bit tricky figuring out how to handle the events and passing data between components. I didn't expect the event to be passed along the way it was, but it did make life easier. Not 100% on how it's doing what it's doing though.